pyproject.toml
[project]
name = "python-notes-organizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.32.0",
    "streamlit-oauth>=0.1.4",
    "google-api-python-client>=2.120.0",
    "google-auth>=2.29.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "google-genai>=1.68.0",
]
